+++ src/unsw/blackout/satellites/SovietSatellite.java
@@ -11,7 +11,7 @@ public class SovietSatellite extends Satellite {
     private int    regionRangeMax     = 190;
 
     // Constructor
-    public SovietSatellite(int height, String id, float position) {
+    public SovietSatellite(int height, String id, double position) {
         super(height, id, position, "SovietSatellite", 6000);
     }
